requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
asn1crypto==0.24.0        # via cryptography
backcall==0.1.0           # via ipython
bcrypt==3.1.4             # via paramiko
cffi==1.11.5              # via bcrypt, cryptography, pynacl
cryptography==2.3.1       # via fabric, paramiko
decorator==4.3.0          # via ipython, traitlets
django-file-resubmit==0.5.1
django-smart-selects==1.5.4
django==2.1.1
djangorestframework==3.8.2
fabric==2.3.1
gunicorn==19.9.0
idna==2.7                 # via cryptography
invoke==1.1.1             # via fabric
ipython-genutils==0.2.0   # via traitlets
ipython==6.5.0
jedi==0.12.1              # via ipython
paramiko==2.4.1           # via fabric
parso==0.3.1              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.4        # via ipython
pillow==5.2.0
prompt-toolkit==1.0.15    # via ipython
ptyprocess==0.6.0         # via pexpect
pyasn1==0.4.4             # via paramiko
pycparser==2.18           # via cffi
pygments==2.2.0
pynacl==1.2.1             # via paramiko
pytz==2018.5              # via django
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via bcrypt, cryptography, prompt-toolkit, pynacl, traitlets
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
